Fix manager usr_group and change name table user_group by usr_group_user

+++ application/controllers/usr_group_controller.php
@@ -48,8 +48,9 @@
         if ($id && !$group->get($id)) {
             redirect($back);
         }
-        
-        $group->delete($group->id);
+        if(!$group->isExitUserByUsrGroup($group->id)){
+            $group->delete($group->id);
+        }
         redirect($back);
     
     }
